mcp-server/requirements.txt
fastmcp
uvicorn
chromadb
sentence-transformers
gitpython
requests
pyyaml
PyPDF2
